src/main/java/com/scarv/starwars/service/PlanetService.java
package com.scarv.starwars.service;

import com.scarv.starwars.domain.Planet;
import com.scarv.starwars.repository.PlanetRepository;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class PlanetService {

  private final PlanetRepository planetRepository;

  public PlanetService(PlanetRepository planetRepository) {
    this.planetRepository = planetRepository;
  }

  public Planet create(Planet planet) {
    return planetRepository.save(planet);
  }

  public Optional<Planet> get(Long id) {
    return planetRepository.findById(id);
  }

  public Optional<Planet> getByName(String name){
    return planetRepository.findByName(name);
  }

}
